fixed cast to float in the script

+++ lab4/run.py
@@ -18,12 +18,12 @@ if __name__ == '__main__':
 	for bin_file in file_list:	
 		with open('{name}-1.csv'.format(name=bin_file), 'w+') as log_file:
 			for pntr in range(n1, n2+1, delta):
-				log_file.write("{pntr},{result}\n".format(pntr=pntr,result=int(commands.getstatusoutput("./{name} {pntr} 1".format(name=bin_file, pntr=pntr))[1].split("\n")[-1])))
+				log_file.write("{pntr},{result}\n".format(pntr=pntr,result=float(commands.getstatusoutput("./{name} {pntr} 1".format(name=bin_file, pntr=pntr))[1].split("\n")[-1])))
 
 				log_file.flush()
 
 		with open('{name}-4.csv'.format(name=bin_file), 'w+') as log_file:
 			for pntr in range(n1, n2+1, delta):
-				log_file.write("{pntr},{result}\n".format(pntr=pntr,result=int(commands.getstatusoutput("./{name} {pntr} 4".format(name=bin_file, pntr=pntr))[1].split("\n")[-1])))
+				log_file.write("{pntr},{result}\n".format(pntr=pntr,result=float(commands.getstatusoutput("./{name} {pntr} 4".format(name=bin_file, pntr=pntr))[1].split("\n")[-1])))
 				log_file.flush()
 	sys.exit(0)
